+++ galaxy/accounts/tests/test_custom_user_model.py
@@ -67,9 +67,13 @@ class CustomUserModelTest(TestCase):
         assert user.email == ""
         assert user.is_active
         assert user.date_joined == NOW
-        assert user.karma == 0
         assert user.avatar_url == ""
-        assert not user.cache_refreshed
+        assert not user.notify_survey
+        assert not user.notify_import_fail
+        assert not user.notify_import_success
+        assert not user.notify_content_release
+        assert not user.notify_author_release
+        assert not user.notify_galaxy_announce
 
         fake_now.assert_called_once()
