+++ galaxy/api/serializers/serializers.py
@@ -39,9 +39,6 @@ from galaxy.main.models import (Platform,
 
 __all__ = [
     'BaseSerializer',
-    'MeSerializer',
-    'UserListSerializer',
-    'UserDetailSerializer',
     'SubscriptionSerializer',
     'StargazerSerializer',
     'CategorySerializer',
@@ -191,168 +188,6 @@ class BaseSerializer(serializers.ModelSerializer):
                 return obj.active
             except AttributeError:
                 return None
-
-
-class MeSerializer(BaseSerializer):
-    authenticated = serializers.SerializerMethodField()
-    staff = serializers.ReadOnlyField(source='is_staff')
-
-    class Meta:
-        model = User
-        fields = ('id', 'authenticated', 'staff', 'username',)
-
-    def get_summary_fields(self, obj):
-        if obj is None or isinstance(obj, AnonymousUser):
-            return {}
-        d = super(MeSerializer, self).get_summary_fields(obj)
-        return d
-
-    def get_authenticated(self, obj):
-        return obj.is_authenticated()
-
-
-class UserListSerializer(BaseSerializer):
-    staff = serializers.ReadOnlyField(source='is_staff')
-
-    class Meta:
-        model = User
-        fields = (
-            'id',
-            'url',
-            'related',
-            'summary_fields',
-            'created',
-            'modified',
-            'username',
-            'staff',
-            'full_name',
-            'date_joined',
-            'avatar_url',
-            'cache_refreshed'
-        )
-
-    def get_related(self, obj):
-        if obj is None or isinstance(obj, AnonymousUser):
-            return {}
-        res = super(UserListSerializer, self).get_related(obj)
-        res.update(dict(
-            subscriptions=reverse(
-                'api:user_subscription_list', args=(obj.pk,)),
-            starred=reverse(
-                'api:user_starred_list', args=(obj.pk,)),
-            repositories=reverse(
-                'api:user_repositories_list', args=(obj.pk,)),
-            token=reverse(
-                'api:user_token_view', args=(obj.pk,)),
-        ))
-        return res
-
-    def get_summary_fields(self, obj):
-        if obj is None or isinstance(obj, AnonymousUser):
-            return {}
-        d = super(UserListSerializer, self).get_summary_fields(obj)
-        d['subscriptions'] = [
-            OrderedDict([
-                ('id', g.id),
-                ('github_user', g.github_user),
-                ('github_repo', g.github_repo)
-            ]) for g in obj.subscriptions.all()]
-        d['starred'] = [
-            OrderedDict([
-                ('id', g.id),
-                ('github_user', g.repository.github_user),
-                ('github_repo', g.repository.github_repo)
-            ]) for g in obj.starred.select_related('repository').all()]
-        return d
-
-
-class UserDetailSerializer(BaseSerializer):
-    password = serializers.CharField(
-        required=False,
-        write_only=True,
-        default='',
-        help_text='Write-only field used to change the password.'
-    )
-    staff = serializers.ReadOnlyField(source='is_staff')
-
-    class Meta:
-        model = User
-        fields = (
-            'id',
-            'url',
-            'related',
-            'summary_fields',
-            'created',
-            'modified',
-            'username',
-            'password',
-            'karma',
-            'staff',
-            'full_name',
-            'date_joined',
-            'avatar_url',
-            'cache_refreshed'
-        )
-
-    def to_native(self, obj):
-        ret = super(UserDetailSerializer, self).to_native(obj)
-        ret.pop('password', None)
-        ret.fields.pop('password', None)
-        return ret
-
-    def get_validation_exclusions(self):
-        ret = super(UserDetailSerializer, self).get_validation_exclusions()
-        ret.append('password')
-        return ret
-
-    def restore_object(self, attrs, instance=None):
-        new_password = attrs.pop('password', None)
-        instance = super(UserDetailSerializer, self).restore_object(
-            attrs, instance)
-        instance._new_password = new_password
-        return instance
-
-    def save_object(self, obj, **kwargs):
-        new_password = getattr(obj, '_new_password', None)
-        if new_password:
-            obj.set_password(new_password)
-        if not obj.password:
-            obj.set_unusable_password()
-        return super(UserDetailSerializer, self).save_object(obj, **kwargs)
-
-    def get_related(self, obj):
-        if obj is None or isinstance(obj, AnonymousUser):
-            return {}
-        res = super(UserDetailSerializer, self).get_related(obj)
-        res.update(dict(
-            repositories=reverse(
-                'api:user_repositories_list', args=(obj.pk,)),
-            subscriptions=reverse(
-                'api:user_subscription_list', args=(obj.pk,)),
-            starred=reverse(
-                'api:user_starred_list', args=(obj.pk,)),
-            token=reverse(
-                'api:user_token_view', args=(obj.pk,)),
-        ))
-        return res
-
-    def get_summary_fields(self, obj):
-        if obj is None or isinstance(obj, AnonymousUser):
-            return {}
-        d = super(UserDetailSerializer, self).get_summary_fields(obj)
-        d['subscriptions'] = [
-            OrderedDict([
-                ('id', g.id),
-                ('github_user', g.github_user),
-                ('github_repo', g.github_repo)
-            ]) for g in obj.subscriptions.all()]
-        d['starred'] = [
-            OrderedDict([
-                ('id', g.id),
-                ('github_user', g.repository.github_user),
-                ('github_repo', g.repository.github_repo)
-            ]) for g in obj.starred.select_related('repository').all()]
-        return d
 
 
 class SubscriptionSerializer(BaseSerializer):
