+++ galaxy/api/views/__init__.py
@@ -28,4 +28,5 @@ from .provider_namespace import *   # noqa
 from .roles import *                # noqa
 from .search import *               # noqa
 from .token import *                # noqa
+from .users import *                # noqa
 from .views import *                # noqa
